requirements.txt
click>=8.0.0
pandas>=1.3.0
pydantic>=2.0.0
hanzipy>=1.0.0
pypinyin>=0.44.0
requests>=2.25.0
playsound3>=3.0.0
ebooklib>=0.18
azure-ai-textanalytics==5.3.0
openai>=1.0.0
google-generativeai>=0.8.0
PyYAML>=6.0.0
dashscope>=1.20.0
pydub>=0.25.0
tencentcloud-sdk-python-tts>=3.0.1427
